pyproject.toml
[tool.poetry]
name = "airbnb-clone-assignment-24"
version = "0.1.0"
description = ""
authors = ["Jay Son <sweetandsourkiss@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
